Fix all clippy lints and apply rustfmt
Newer toolchain lints: box Value::Recur (the variant dominated the enum
at ~312 bytes vs 24 for the next largest), add type aliases for the
parse return type in the bindings and the ical-crate test helper, and
apply the mechanical clippy --fix suggestions and rustfmt drift.

+++ crates/vobject-core/src/value/datetime.rs
@@ -192,7 +192,7 @@ impl Weekday {
 
     /// Days from `self` to `other` going forward (0-6).
     pub fn days_until(&self, other: Weekday) -> u8 {
-        ((other.number0() + 7 - self.number0()) % 7) as u8
+        (other.number0() + 7 - self.number0()) % 7
     }
 
     /// Monday = 0 ... Sunday = 6.
@@ -411,10 +411,22 @@ mod tests {
 
     #[test]
     fn weekday_computation() {
-        assert_eq!(Date::parse("20260722").unwrap().weekday(), Weekday::Wednesday);
-        assert_eq!(Date::parse("20000101").unwrap().weekday(), Weekday::Saturday);
-        assert_eq!(Date::parse("19700101").unwrap().weekday(), Weekday::Thursday);
-        assert_eq!(Date::parse("20240229").unwrap().weekday(), Weekday::Thursday);
+        assert_eq!(
+            Date::parse("20260722").unwrap().weekday(),
+            Weekday::Wednesday
+        );
+        assert_eq!(
+            Date::parse("20000101").unwrap().weekday(),
+            Weekday::Saturday
+        );
+        assert_eq!(
+            Date::parse("19700101").unwrap().weekday(),
+            Weekday::Thursday
+        );
+        assert_eq!(
+            Date::parse("20240229").unwrap().weekday(),
+            Weekday::Thursday
+        );
     }
 
     #[test]
